// src/modules/people/types.ts
type male = 'male' | 'female'

export interface Person {
  name: string
  height: string
  mass: string
  hair_color: string
  skin_color: string
  eye_color: string
  birth_year: string
  gender: male
  homeworld: string
  films: string[]
  species: string[]
  vehicles: string[]
  starships: string[]
  created: string
  edited: string
  url: string
}

export interface PeopleInitialState {
  data: Person[]
  isFetching: boolean
  isFetched: boolean
}